results/rebuttal/run-20231111-034650_LLMatic-zero_shot_donnemartin-system-design-oop-bookclub_0_20-runs_log/bookclub/16/LLMatic/test_app.py
import app

def test_database():
	# Test that the database can store and retrieve data
	app.DATABASE['users'] = {}
	app.DATABASE['users']['test_user'] = app.User('test_user', 'test_password')
	assert isinstance(app.DATABASE['users']['test_user'], app.User)

def test_user_creation():
	# Test user creation
	app.DATABASE['users'] = {}
	response = app.User.create_user('test_user', 'test_password')
	assert response == 'User created successfully'
	assert isinstance(app.DATABASE['users']['test_user'], app.User)

def test_user_authentication():
	# Test user authentication
	app.DATABASE['users'] = {}
	app.User.create_user('test_user', 'test_password')
	response = app.User.authenticate_user('test_user', 'test_password')
	assert response == 'User authenticated'
	response = app.User.authenticate_user('test_user', 'wrong_password')
	assert response == 'Authentication failed'

def test_get_user():
	# Test get user
	app.DATABASE['users'] = {}
	app.User.create_user('test_user', 'test_password')
	user = app.User.get_user('test_user')
	assert isinstance(user, app.User)
	assert user.username == 'test_user'
	assert user.password == 'test_password'

def test_book_club_creation():
	# Test book club creation
	app.DATABASE['book_clubs'] = {}
	response = app.BookClub.create_book_club('test_club', False)
	assert response == 'Book club created successfully'
	assert isinstance(app.DATABASE['book_clubs']['test_club'], app.BookClub)

def test_get_book_club():
	# Test get book club
	app.DATABASE['book_clubs'] = {}
	app.BookClub.create_book_club('test_club', False)
	book_club = app.BookClub.get_book_club('test_club')
	assert isinstance(book_club, app.BookClub)
	assert book_club.name == 'test_club'
	assert book_club.is_private == False

def test_add_member():
	# Test add member
	app.DATABASE['book_clubs'] = {}
	app.BookClub.create_book_club('test_club', False)
	book_club = app.BookClub.get_book_club('test_club')
	response = book_club.add_member('test_user')
	assert response == 'User added to the book club'
	assert 'test_user' in book_club.members

def test_remove_member():
	# Test remove member
	app.DATABASE['book_clubs'] = {}
	app.BookClub.create_book_club('test_club', False)
	book_club = app.BookClub.get_book_club('test_club')
	book_club.add_member('test_user')
	response = book_club.remove_member('test_user')
	assert response == 'User removed from the book club'
	assert 'test_user' not in book_club.members

def test_schedule_meeting():
	# Test schedule meeting
	app.DATABASE['meetings'] = {}
	response = app.Meeting.schedule_meeting('test_meeting', 'test_club', '2022-12-31', '12:00')
	assert response == 'Meeting scheduled successfully'
	assert isinstance(app.DATABASE['meetings']['test_meeting'], app.Meeting)

def test_get_meeting():
	# Test get meeting
	app.DATABASE['meetings'] = {}
	app.Meeting.schedule_meeting('test_meeting', 'test_club', '2022-12-31', '12:00')
	meeting = app.Meeting.get_meeting('test_meeting')
	assert isinstance(meeting, app.Meeting)
	assert meeting.meeting_id == 'test_meeting'
	assert meeting.book_club_name == 'test_club'
	assert meeting.date == '2022-12-31'
	assert meeting.time == '12:00'
